+++ docs/changelog.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - Change single logger to a logging hierarchy
  - Decompose models
- - Optimize ScTemplate and change list on tuple in params
+ - Optimize ScTemplate, ScTemplateResult and change list to tuple in params
 ### Deprecated
  - ScModule, ScAgent and ScKeynode implementations moved to [py-sc-kpm](https://github.com/ostis-ai/py-sc-kpm)